【RE】[東京大学プログラミングコンテスト2012]C - 森ですか？

+++ utpc2012/utpc2012_03.cpp
@@ -52,7 +52,7 @@ P e(P p) {
 }
 
 int main() {
-    int n, m;
+    ll n, m;
     cin >> n >> m;
 
     vector<P> v(m);
